resources/Pomodoro Github Clone/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = ""
authors = [
    {name = "Rukenshia",email = "code@chrstphrsn.cc"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow (>=11.1.0,<12.0.0)"
]

